Add XUDP support by simply renaming vudp to xudp
https://t.me/projectXray/243505
