+++ common/xudp/xudp.go
@@ -1,4 +1,4 @@
-package vudp
+package xudp
 
 import (
 	"io"
